Refactor: Change directory structure to add 404 page #36
