--- src/app/(main)/not-found.module.css
@@ -1,3 +1,0 @@
-.notFound{
-    
-}
