+++ src/app/(onboard)/layout.js
@@ -1,15 +1,11 @@
-import "@/style/global.css";
+import '@/style/global.css'
 
 export const metadata = {
-  title: "ATC 2024 온보딩 페이지",
+  title: 'ATC 2024 온보딩 페이지',
   description:
-    "코끼리를 냉장고에 넣는 방법, 2024년 Art&Technolgoy 컨퍼런스 웹의 온보딩 페이지",
-};
+    '코끼리를 냉장고에 넣는 방법, 2024년 Art&Technolgoy 컨퍼런스 웹의 온보딩 페이지'
+}
 
 export default function OnBoardRootLayout({ children }) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  );
+  return <>{children}</>
 }
