+++ src/app/not-found.module.css
@@ -1,0 +1,59 @@
+.notFound {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.notFound div {
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.notFound div h1 {
+  color: #000;
+  font-family: 'Sandoll Seoul';
+  font-size: 30rem;
+  font-style: normal;
+  line-height: 50%;
+  font-weight: 400;
+  letter-spacing: 6rem;
+
+  @media (max-width: 768px) {
+    font-size: 10rem;
+    letter-spacing: 3rem;
+  }
+}
+
+.notFound div p {
+  color: #000;
+  font-family: 'Sandoll Seoul';
+  font-size: 4rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 150%; /* 6rem */
+  letter-spacing: 0.8rem;
+  @media (max-width: 768px) {
+    font-size: 3rem;
+  }
+}
+
+.notFound div button {
+  margin-top: 9.5rem;
+  color: #fff;
+  text-align: center;
+  font-family: Pretendard;
+  font-size: 2rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 150%; /* 3rem */
+  letter-spacing: 0.1rem;
+  background-color: black;
+  padding: 1rem 3rem;
+  @media (max-width: 768px) {
+    margin-top: 3rem;
+  }
+}
